Move Server unit tests that require rendering to a separate file.
